Improved testing. added component tests on selectorscreen and on the CustomListEmptyComponent inside of it
