+++ package.json
@@ -6,7 +6,7 @@
     "android": "react-native run-android",
     "ios": "react-native run-ios",
     "start": "react-native start",
-    "test": "jest",
+    "test": "jest --silent",
     "lint": "eslint .",
     "docs": "jsdoc -c jsdoc.conf.json",
     "prepare": "patch-package"
